requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
groq==0.4.1
httpx==0.25.2
ruff==0.1.6
mypy==1.7.1